+++ lesson17_map_filter_lambda/task3_is_even_lambda.py
